openstack: rename prefix for resources from fstests to kdevops
The fstests name comes from the fact our lineage of code traces back
to when this code was oscheck [0] specific with the only intent on using
the workflow to test fstests [1]. After it became clear that the ramp up
and provisioning work was generic kdevops was born as an independent
project which oscheck could instead rely on. So just change the prefix
name to remove the old provenance name stuff.

This goes tested against minicloud.

[0] https://github.com/mcgrof/oscheck
[1] https://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git/

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>