aws: replace credential as variables to shared credentials file
We currently login using an access key and secret key but these
are required to be in your terraform.tfvars file, this is risky
for some projects as it may mean you have to expose the liability
of a project not adding the terraform.tfvars file to .gitignore.

This is specially true now that this is an ansible role, we have
no control over our users. AWS supports a shared credentials file
which you can have as by defulat:

  ~/.aws/credentials

This file is rather simple with a structure as follows:

[default]
aws_access_key_id = SOME_ACCESS_KEY
aws_secret_access_key = SECRET_KEY

And, you can have multiple entries, in this case the default
profile is "default", but this also lets you have credentials
for different logins in one file and you can just specifify
now the profile for access in your terraform.tfvars.

Since we default to this now, all you really need is just to
specify your profile, if it differs from "default". The region
is always required.

Note that since we are removing support for the old authentication
mechanism, if you were using this role before you have to upgrade
to the new scheme, otherwise things won't work. I'm OK with that
as we're not yet in 1.0.0 release state.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>