all: move ssh_config_user to be the variable for username
Beacuse of Azure, on which bring up doesn't not gaurantee that
the system is yet accessible we have had to hold back on setting
the ssh defaults to *not* update your ~/.ssh/config for you.

This however is just silly. Document that the user has to set
this to false then, and in fact, when we add kconfig support
for terraform, we just default it to false there.

This way we don't hold back other cloud solutions.

This consolidates the username definition in once place, to avoid
confusion. I can't find out why I hadn't joined these before, but
the only thing I can think of was that we didn'd this yet
because Azure we still pending.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>