aws: fix drive ordering
The aws_ebs_device_names is a complete utter mess.
For instance when Amazon uses PV guests a request for
an EBS volume using /dev/sdh as a device name will actually
end up as /dev/xvdh. Then for KVM guests which support
nvme, this will end up as an nvme drive after the instance
type device names which are supported.

Document some of this mess and fix the use of the
aws_ebs_device_names naming so that we ensure folks demoing
this code will end up with matching /dev/xvdh on 2 nodes
if 2 nodes created.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>