aws: Adjust sizes for EBS to match between 2 default hosts
The sizes were incorrectly used unde the assumption one would
be used for Linux kernel development and the other as a scratch
for fstests testing. But the first one is just the first EBS
for the fist host, and the second one listed the second size
used for the second host.

Thinking about this and the EBS naming mess more, the EBS
naming / specification is a total cluster fuck. And so let's
just disable it for now as a default measure.

Use an instance type with the drives you need for now.

Once an eager bever finds a good need for this and figures
out a nice way to specify number of drives per host and
in an order which makes sense, we can enable them. But
in retrospect this was just a complete utter mess.

Let's recommend simple solutios without EBS for now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>