azure: use static IP assignement as dynamic is broken
Azure's dynamic IP address assignment doesn't exactly work
too well, but it kind of makes sense if you think about it.

There is a difference between using a dynamic IP address *once*
and using a static IP address... without specification.

Now, it turns out that the IP address assignment for the type
"Static" does not require you as a user to specify an IP
address. So actually, from a user perspective, this IP address
is also dynamic. And that works on Azure, but it still requires
a "depends on" logic which we correct and fix so that the machine
is up first before printing the IP address out. And that's fine.

However.. it remains unclear if the issue is that Microsoft Azure
team thinks that the type "Dynamic" means you should get a *new*
IP *every time* you boot up. If so, that would explain the issue...

But even so, the current print out of the dynamic IP address
doesn't even print out anything, regardless of how hard you work
on creating dependencies on the IP print out.

And so I give up on Azure dynamic IP address use. Just don't
use it.

So we default to static IP address assigment, even though we
don't specify an address.

Details of this issue are specified here:

https://github.com/terraform-providers/terraform-provider-azurerm/issues/310

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>