gce: add additional scratch disk
This is added to match our defaults for kdevops.
Ideally later we want to get the disks we need from
the yaml file but this proves a bit difficult at the time
given the diversity in which each cloud provider specifies
how to create additional drives *and* the interfaces used
vary considerably. For instance Azure is stuck in the stone
age and uses scsi because their hypervisor uses a scsi-front
end driver for the hypervisor.

We can revisit this later.

For now I'm addign enough defautls to get most kernel
developmet going.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>