update_ssh_config: update to v2.2.0 to do updates in one shot
Now that the Python script is being well tested there is high
confidence in the code, so no need to do two separate operations.
This also lets us share the strategy with vagrant, and the only
difference being the final name extension, indicating the method
used to do the bring up: vagrant or terraform.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>